+++ server/providers.go
@@ -1,42 +1,47 @@
 package server
 
 import (
-	"github.com/vseinstrumentiru/lego/v2/internal/metrics"
-	"github.com/vseinstrumentiru/lego/v2/internal/metrics/exporters/newrelicexporter"
-	"github.com/vseinstrumentiru/lego/v2/internal/metrics/propagation"
-	events2 "github.com/vseinstrumentiru/lego/v2/transport/events"
-	eventrouter2 "github.com/vseinstrumentiru/lego/v2/transport/events/eventrouter"
+	"github.com/vseinstrumentiru/lego/v2/metrics"
+	"github.com/vseinstrumentiru/lego/v2/metrics/exporters/newrelicexporter"
+	"github.com/vseinstrumentiru/lego/v2/metrics/propagation"
+	"github.com/vseinstrumentiru/lego/v2/transport/events"
+	"github.com/vseinstrumentiru/lego/v2/transport/events/eventrouter"
 	"github.com/vseinstrumentiru/lego/v2/transport/grpc"
 	"github.com/vseinstrumentiru/lego/v2/transport/http/httpclient"
 	"github.com/vseinstrumentiru/lego/v2/transport/http/httpserver"
-	mysql2 "github.com/vseinstrumentiru/lego/v2/transport/mysql"
-	nats2 "github.com/vseinstrumentiru/lego/v2/transport/nats"
+	"github.com/vseinstrumentiru/lego/v2/transport/mysql"
+	"github.com/vseinstrumentiru/lego/v2/transport/nats"
 	"github.com/vseinstrumentiru/lego/v2/transport/sql"
-	stan2 "github.com/vseinstrumentiru/lego/v2/transport/stan"
+	"github.com/vseinstrumentiru/lego/v2/transport/stan"
 )
 
-func providers() []interface{} {
+func providers(runtime *Runtime) []interface{} {
+	if runtime.Is(optWithoutProviders) {
+		return []interface{}{}
+	}
+
 	return []interface{}{
 		// metric providers
 		propagation.ProvideHTTP,
 		newrelicexporter.Provide,
-		metrics.Provide,
+		metrics.ProvideHealthChecker,
+		metrics.ProvideMonitoringServer,
 		// http / grpc
 		httpserver.Provide,
 		httpclient.Provide,
 		httpclient.ConstructorProvider,
 		grpc.Provide,
 		// database
-		mysql2.Provide,
+		mysql.Provide,
 		sql.Provide,
 		// events
-		nats2.Provide,
-		stan2.Provide,
-		eventrouter2.Provide,
-		events2.ProvideKafkaPublisher,
-		events2.ProvideKafkaSubscriber,
-		events2.ProvideNatsSubscriber,
-		events2.ProvideNatsPublisher,
-		events2.ProvideChannel,
+		nats.Provide,
+		stan.Provide,
+		eventrouter.Provide,
+		events.ProvideKafkaPublisher,
+		events.ProvideKafkaSubscriber,
+		events.ProvideNatsSubscriber,
+		events.ProvideNatsPublisher,
+		events.ProvideChannel,
 	}
 }